refactor(milestone): inline milestone design config
