+++ packages/design/smb/index.ts
@@ -1,10 +1,7 @@
-import Milestone from './src/milestone'
 import { version } from './package.json'
 
 export default {
   name: 'smb',
   version,
-  components: {
-    Milestone
-  }
+  components: {}
 }
